Update audit_logger, config_manager, dashboard_server, enhanced_dashboard, enhanced_scanner, and requirements.txt

+++ enhanced_scanner.py
@@ -526,9 +526,9 @@ class ContinuousMarketScanner:
     
     def _calculate_rsi(self, prices: pd.Series, period: int = 14) -> pd.Series:
         """Calculate RSI"""
-        delta = prices.diff()
-        gain = (delta.where(delta > 0, 0)).rolling(window=period).mean()
-        loss = (-delta.where(delta < 0, 0)).rolling(window=period).mean()
+        delta = prices.diff().astype(float)
+        gain = delta.where(delta > 0, 0.0).rolling(window=period).mean()
+        loss = (-delta.where(delta < 0, 0.0)).rolling(window=period).mean()
         rs = gain / loss
         return 100 - (100 / (1 + rs))
